stable improvement

+++ web-ordering/src/components/CartComponent.js
@@ -171,25 +171,26 @@ const Cart = (props) => {
       querySnapshot.forEach(function (doc) {
         var nn = doc.id;
         koiSushiRestaurant.collection("tables").doc(tablenumber).collection("cart").doc(doc.id).update({ ID: nn });
-        if (dic.has(doc.data().dishRef.id.path)) {
+        console.log(doc.data().dishRef);
+        if (dic.has(doc.data().dishRef.ID)) {
           qiangbimingdan.add(doc.data().ID);
-          var right = dic.get(doc.data().dishRef.id.path);
-          if (right != null) {
-            koiSushiRestaurant.collection("tables").doc(tablenumber).collection("cart").doc(right).get().then(function (dddoc) {
-              oldnumber = dddoc.data().number;
-              console.log("dddd!" + oldnumber);
-              handleAdd(parseInt(oldnumber), right);
-            })
-          }
+          // var right = dic.get(doc.data().dishRef.ID);
+          // if (right != null) {
+          //   koiSushiRestaurant.collection("tables").doc(tablenumber).collection("cart").doc(right).get().then(function (dddoc) {
+          //     oldnumber = dddoc.data().number;
+          //     console.log("dddd!" + oldnumber);
+          //     handleAdd(parseInt(oldnumber), right);
+          //   })
+          // }
         } else {
-          dic.set(doc.data().dishRef.id.path, doc.data().ID);
+          dic.set(doc.data().dishRef.ID, doc.data().ID);
           // console.log("put!");
         }
-        qiangbimingdan.forEach(function (kkk) {
-          if (kkk != null) {
-            koiSushiRestaurant.collection("tables").doc(tablenumber).collection("cart").doc(kkk).delete();
-          }
-        })
+        // qiangbimingdan.forEach(function (kkk) {
+        //   if (kkk != null) {
+        //     koiSushiRestaurant.collection("tables").doc(tablenumber).collection("cart").doc(kkk).delete();
+        //   }
+        // })
       });
       setFinishFilter(true);
     })
@@ -204,7 +205,22 @@ const Cart = (props) => {
     if (newnum > 1) {
       koiSushiRestaurant.collection("tables").doc(tablenumber).collection("cart").doc(idid).update({ number: newnum - 1 });
     } else {
-      koiSushiRestaurant.collection("tables").doc(tablenumber).collection("cart").doc(idid).delete();
+      idid = String(idid);
+      var sk;
+      koiSushiRestaurant.collection("tables").doc(tablenumber).collection("cart").get().then(snap => {
+        // sk = snap.size // will return the collection size
+        console.log(parseInt(snap.size));
+        if(snap.size === 1){
+          handleClear(testcart);
+        }else{
+          koiSushiRestaurant.collection("tables").doc(tablenumber).collection("cart").doc(idid).delete();
+        }
+      });
+      // console.log(parseInt(sk));
+
+
+      updatepc();
+
     }
     updatepc();
